#14610: Split routing_info_t out of dev_msgs.h
