+++ tt_metal/hw/inc/routing_info.h
@@ -1,0 +1,20 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+struct eth_word_t {
+    volatile uint32_t bytes_sent;
+    volatile uint32_t dst_cmd_valid;
+    uint32_t reserved_0;
+    uint32_t reserved_1;
+};
+
+struct routing_info_t {
+    volatile uint32_t routing_enabled;
+    volatile uint32_t src_sent_valid_cmd;
+    volatile uint32_t dst_acked_valid_cmd;
+    volatile uint32_t unused_arg0;
+    eth_word_t fd_buffer_msgs[2];
+};
